Rename AI/Trait module to AI/Character for review #677.
