+++ default/python/tests/AI/Character/freeOrionAIInterface.py
